1.12 nightmare on elm street
I understand why they are implementing more and more json, it's part of
the grand unification of the clients and servers things. I saw some of
the plans for 1.13, the goal is to have servers to contain all of the
data for a mod. The client to contain none of it. The client will
download from the server it connects to all of the relevant data,
textures, functions etc. Essentially this means that upon a particular
release I would assume that to be 2.0 a console would be able to connect
to a forged server. Thus allowing everyone to play together. I saw
concern about JEB's comment about the java version to be considered the
experimental playground. Some on youtube were very concerned about that
statement. I am not. All of the Linux distros have both a commerical and
community (expiremental) counterpart. The community is where things are
honed out. I'm happy to be on bleeding edge. :)